HWTS-72: makefile to automate hashsums update
Discussing a potential strategy for using make to automate
the updating of a hashsum file. This is for the HOWTO
'Recursive sha256 workflow'. Discuss drawbacks for this approach.

Also mention a Bash script I am currently using which
works for me for this purpose.

Implements https://github.com/valera-rozuvan/howtos/issues/72